+++ source/slang/slang-serialize-ast.h
@@ -30,6 +30,7 @@ class ModuleSerialFilter : public SerialFilter
 public:
     // SerialFilter impl
     virtual SerialIndex writePointer(SerialWriter* writer, const NodeBase* ptr) SLANG_OVERRIDE;
+    virtual SerialIndex writePointer(SerialWriter* writer, const RefObject* ptr) SLANG_OVERRIDE;
 
     ModuleSerialFilter(ModuleDecl* moduleDecl):
         m_moduleDecl(moduleDecl)
@@ -39,34 +40,10 @@ public:
     ModuleDecl* m_moduleDecl;
 };
 
-class DefaultSerialObjectFactory : public SerialObjectFactory
+struct ASTSerialUtil
 {
-public:
-
-    virtual void* create(SerialTypeKind typeKind, SerialSubType subType) SLANG_OVERRIDE;
-
-    DefaultSerialObjectFactory(ASTBuilder* astBuilder) :
-        m_astBuilder(astBuilder)
-    {
-    }
-
-protected:
-    RefObject* _add(RefObject* obj)
-    {
-        m_scope.add(obj);
-        return obj;
-    }
-
-    // We keep RefObjects in scope 
-    List<RefPtr<RefObject>> m_scope;
-    ASTBuilder* m_astBuilder;
-};
-
-/* None of the functions in this util should *not* be called from production code,
-they exist to test features of AST Serialization */
-struct ASTSerialTestUtil
-{
-    static SlangResult selfTest();
+        /// Add the AST related classes
+    static void addSerialClasses(SerialClasses* classes);
 
         /// Tries to serialize out, read back in and test the results are the same.
         /// Will write dumped out node to files
