+++ source/slang/slang.vcxproj
@@ -286,8 +286,10 @@
     <ClInclude Include="slang-ref-object-reflect.h" />
     <ClInclude Include="slang-reflection.h" />
     <ClInclude Include="slang-repro.h" />
+    <ClInclude Include="slang-serialize-ast-type-info.h" />
     <ClInclude Include="slang-serialize-ast.h" />
     <ClInclude Include="slang-serialize-container.h" />
+    <ClInclude Include="slang-serialize-factory.h" />
     <ClInclude Include="slang-serialize-ir-types.h" />
     <ClInclude Include="slang-serialize-ir.h" />
     <ClInclude Include="slang-serialize-misc-type-info.h" />
@@ -406,6 +408,7 @@
     <ClCompile Include="slang-repro.cpp" />
     <ClCompile Include="slang-serialize-ast.cpp" />
     <ClCompile Include="slang-serialize-container.cpp" />
+    <ClCompile Include="slang-serialize-factory.cpp" />
     <ClCompile Include="slang-serialize-ir-types.cpp" />
     <ClCompile Include="slang-serialize-ir.cpp" />
     <ClCompile Include="slang-serialize-reflection.cpp" />
